fix: lint errors

+++ src/math/comb/binom_fn.hpp
@@ -2,7 +2,7 @@
 #include <cstdint>
 
 /// nCk; $O(k)$
-inline int64_t binom(int64_t n, int k) {
+inline int64_t binom(int64_t n, int64_t k) {
   if (k < 0 || n < k) return 0;
   if (k > n / 2) k = n - k;
   int64_t numer = 1, denom = 1;
